user authentication

+++ controllers/books.controller.js
@@ -28,13 +28,15 @@ const getBook = async(req, res) => {
 const addBook = async(req, res) => {
     try {
         const author = await Author.findOne({
-            where: req.body.author
+            where: {
+                id:req.user.id
+            }
         })
 
         if (!author) {return res.status(500).json({body:"Author was not found"})}
 
-        req.body.book["author"] = author.id
-        const newBook = await Book.create(req.body.book)
+        req.body["bookAuthor"] = author.id
+        const newBook = await Book.create(req.body)
 
         res.status(201).json({body:newBook})
     } catch (err) {
@@ -44,6 +46,9 @@ const addBook = async(req, res) => {
 
 const deleteBook = async(req, res) => {
     try {
+        const book = await Book.findOne({where: req.body})
+        if (book.bookAuthor != req.user.id) {return res.status(403).json({body:"Improper Authorisation"})}
+
         await Book.destroy({where: req.body})
         res.status(200).json({body:true})
     } catch (err) {
@@ -53,8 +58,11 @@ const deleteBook = async(req, res) => {
 
 const updateBook = async(req, res) => {
     try {
-        const book = await Book.update(req.body.newDetails, {where:req.body.currentBook})
-        res.status(200).json({body:book})
+        const book = await Book.findOne({where: req.body.oldDetails})
+        if (book.bookAuthor != req.user.id) {return res.status(403).json({body:"Improper Authorisation"})}
+
+        const bookToReturn = await Book.update(req.body.newDetails, {where:req.body.oldDetails})
+        res.status(200).json({body:bookToReturn})
     } catch (err) {
         res.status(500).json({body:err.message})
     }
